@@ -4,7 +4,7 @@ author: Flutter Authors <flutter-dev@googlegroups.com>
 homepage: http://flutter.dev
 
 environment:
-  sdk: ">=2.0.0-dev.28.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   flutter:
