@@ -2,8 +2,6 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// @dart = 2.8
-
 @TestOn('chrome') // Uses web-only Flutter SDK
 
 import 'dart:ui' as ui; // ignore: unused_import, it looks unused as web-only elements are the only elements used.
@@ -56,8 +54,8 @@ void main() {
       const StandardMessageCodec codec = StandardMessageCodec();
 
       final List<String> loggedMessages = <String>[];
-      ServicesBinding.instance.defaultBinaryMessenger
-          .setMessageHandler('test_send', (ByteData data) {
+      ServicesBinding.instance!.defaultBinaryMessenger
+          .setMessageHandler('test_send', (ByteData? data) {
         loggedMessages.add(codec.decodeMessage(data) as String);
         return null;
       });
@@ -70,14 +68,14 @@ void main() {
           'test_send', codec.encodeMessage('world'));
       expect(loggedMessages, equals(<String>['hello', 'world']));
 
-      ServicesBinding.instance.defaultBinaryMessenger
+      ServicesBinding.instance!.defaultBinaryMessenger
           .setMessageHandler('test_send', null);
     });
 
     test('throws when trying to set a mock handler', () {
       expect(
           () => pluginBinaryMessenger.setMockMessageHandler(
-              'test', (ByteData data) async => ByteData(0)),
+              'test', (ByteData? data) async => ByteData(0)),
           throwsFlutterError);
     });
   });
